Safari-609.1.11.1/LayoutTests/fast/events/popup-blocking-timers1-expected.txt

Click Here
Test calling window.open() directly. A popup should be allowed.
PASS newWindow is non-null.
